Castling!
Castling implemented. Changes to chessPanel, chessModel, and King to properly implement. Done by if castling, then move king and rook to their new positions. When undoing, will check for two moves from the same player in a row, (since this only happens in castling), and check (so you can't castle if it puts you in check, and will properly undo both the king and the rook move).